requirements.txt
transformers>=4.40.0
torch>=2.0.0
datasets>=2.14.0
accelerate>=0.27.0
peft>=0.10.0
bitsandbytes>=0.42.0
trl>=0.8.0
groq
google-generativeai
python-dotenv
requests
sentencepiece
protobuf
scipy